+++ search-films/src/components/Menu/index.tsx
@@ -1,5 +1,7 @@
 import { useContext } from 'react';
+import { useSelector } from 'react-redux';
 import { IUserContext, UserContext } from '../../context/user.context';
+import { RootState } from '../../store/store';
 import LinkedMenuItem from '../LinkedMenuItem';
 import MenuIconBuilder from '../shared/MenuIconBuilder';
 import SimpleMenuItem from '../SimpleMenuItem';
@@ -7,6 +9,8 @@ import styles from './index.module.css';
 
 function Menu() {
 	const {currentUser, isAuthenticated, logoutCurrentUser} = useContext(UserContext) as IUserContext;
+	const favors = useSelector((f: RootState) => f.favors.favorites);
+
 	const authMenu = (
 		<>
 			<LinkedMenuItem
@@ -15,7 +19,7 @@ function Menu() {
 			/>
 			<LinkedMenuItem
 				text="Мои фильмы"
-				icon={MenuIconBuilder.buildCounterIcon(3)}
+				icon={MenuIconBuilder.buildCounterIcon(favors.length)}
 				linkUrl="/favorites"
 				handleClicks={(event) => console.log(event)}
 			/>
@@ -38,7 +42,7 @@ function Menu() {
 			/>
 			<LinkedMenuItem
 				text="Мои фильмы"
-				icon={MenuIconBuilder.buildCounterIcon(3)}
+				icon={MenuIconBuilder.buildCounterIcon(0)}
 				linkUrl="/favorites"
 				handleClicks={(event) => console.log(event)}
 			/>